+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
 - 
 ### Fixed:
-- 
+- Fixed piglin flag render light
 ### Changed:
-- Updated VoidZ dependency
+-